# swagger.yml
openapi: 3.0.0
paths:
...
  /api/categories:
    get:
      operationId: CategoriesController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: limit
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: search
          required: false
          in: query
          description: Search query
          schema:
            type: string
        - name: sortBy
          required: false
          in: query
          schema:
            $ref: '#/components/schemas/Object'
        - name: sortOrder
          required: false
          in: query
          schema:
            default: desc
            type: string
            enum:
              - asc
              - desc
        - name: filter
          required: false
          in: query
          description: Generic filter string
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: &ref_19
        - bearer: []
      summary: List categories with pagination/filter/sort
      tags: &ref_20
        - categories
    post:
      operationId: CategoriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: ''
      security: *ref_19
      tags: *ref_20
...
  /api/categories/{categoryId}:
    put:
      operationId: CategoriesController_replace
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
    patch:
      operationId: CategoriesController_patch
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
    get:
      operationId: CategoriesController_get
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
    delete:
      operationId: CategoriesController_remove
      parameters:
        - name: categoryId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_19
      tags: *ref_20
...
  /api/website/products/{id}:
    get:
      operationId: WebsiteController_getProduct
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get product detail
      tags: *ref_45
...
  /api/website/categories/{slug}:
    get:
      operationId: WebsiteController_getCategoryBySlug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get category by slug with products
      tags: *ref_45
...
  /api/website/prescriptions:
    post:
      operationId: WebsiteController_createPrescription
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreatePrescriptionDto'
      responses:
        '201':
          description: ''
      summary: Upload prescription with files
      tags: *ref_45
    get:
      operationId: WebsiteController_listPrescriptions
      parameters: []
      responses:
        '200':
          description: ''
      summary: List prescriptions
      tags: *ref_45
  /api/website/consultations:
    post:
      operationId: WebsiteController_createConsultation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsultationDto'
      responses:
        '201':
          description: ''
      summary: Book a consultation
      tags: *ref_45
    get:
      operationId: WebsiteController_listConsultations
      parameters: []
      responses:
        '200':
          description: ''
      summary: List consultations
      tags: *ref_45
  /api/website/cart:
    post:
      operationId: WebsiteController_addToCart
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AddToCartDto'
      responses:
        '201':
          description: ''
      summary: Add to cart (returns product details for given ids)
      tags: *ref_45
    get:
      operationId: WebsiteController_getCart
      parameters:
        - name: ids
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get cart product details
      tags: *ref_45
...
  /api/website/orders:
    post:
      operationId: WebsiteController_createOrder
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: ''
      summary: Create order
      tags: *ref_45
    get:
      operationId: WebsiteController_listOrders
      parameters: []
      responses:
        '200':
          description: ''
      summary: List orders
      tags: *ref_45
  /api/website/orders/track/{trackingCode}:
    get:
      operationId: WebsiteController_trackOrder
      parameters:
        - name: trackingCode
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Track order by code
      tags: *ref_45
  /api/website/orders/{id}:
    get:
      operationId: WebsiteController_getOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get order detail
      tags: *ref_45
...
  /api/website/articles/{slug}:
    get:
      operationId: WebsiteController_getArticleBySlug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get article by slug
      tags: *ref_45
...
  /api/website/branches/{id}:
    get:
      operationId: WebsiteController_getBranch
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get branch detail
      tags: *ref_45
  /api/website/contact:
    post:
      operationId: WebsiteController_submitContact
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactDto'
      responses:
        '201':
          description: ''
      summary: Submit contact form
      tags: *ref_45
  /api/website/newsletter/subscribe:
    post:
      operationId: WebsiteController_subscribe
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsletterSubscribeDto'
      responses:
        '201':
          description: ''
      summary: Subscribe to newsletter
      tags: *ref_45
  /api/website/reviews:
    post:
      operationId: WebsiteController_createReview
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewDto'
      responses:
        '201':
          description: ''
      summary: Submit product review
      tags: *ref_45
  /api/website/reviews/{productId}:
    get:
      operationId: WebsiteController_getProductReviews
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get product reviews
      tags: *ref_45
  /api/website/rewards:
    get:
      operationId: WebsiteController_getRewards
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: Get rewards data
      tags: *ref_45
  /api/website/search:
    get:
      operationId: WebsiteController_search
      parameters:
        - name: q
          required: true
          in: query
          schema:
            type: string
        - name: type
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Global search
      tags: *ref_45
...
info:
  title: RxSoft Backend API
  description: Phase 3 identity module and API contracts
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Object:
      type: object
      properties: {}
...
    CreateCategoryDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        parentId:
          type: string
      required:
        - code
        - name
    UpdateCategoryDto:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        parentId:
          type: string
...
    CreatePrescriptionDto:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        notes:
          type: string
    CreateConsultationDto:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        symptoms:
          type: string
        questions:
          type: string
        channel:
          type: string
      required:
        - name
        - phone
    AddToCartDto:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: number
      required:
        - productId
        - quantity
    CreateOrderDto:
      type: object
      properties:
        deliveryAddress:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        shippingMethod:
          type: string
        paymentMethod:
          type: string
        prescriptionIds:
          type: array
          items:
            type: string
        notes:
          type: string
      required:
        - deliveryAddress
        - paymentMethod
    CreateContactDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - subject
        - message
    NewsletterSubscribeDto:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
      required:
        - email
    CreateReviewDto:
      type: object
      properties:
        productId:
          type: string
        rating:
          type: number
        comment:
          type: string
        imageUrls:
          type: array
          items:
            type: string
      required:
        - productId
        - rating
...